+++ polaris-react/src/components/Bleed/Bleed.tsx
@@ -5,28 +5,27 @@ import {sanitizeCustomProperties} from '../../utilities/css';
 
 import styles from './Bleed.scss';
 
-interface Spacing {
-  bottom: SpacingSpaceScale;
-  left: SpacingSpaceScale;
-  right: SpacingSpaceScale;
-  top: SpacingSpaceScale;
-}
-
 export interface BleedProps {
   /** Elements to display inside tile */
   children: React.ReactNode;
-  spacing?: SpacingSpaceScale;
+  /** Negative horizontal space around the element
+   * * @default '5'
+   */
   horizontal?: SpacingSpaceScale;
+  /** Negative vertical space around the element */
   vertical?: SpacingSpaceScale;
+  /** Negative top space around the element */
   top?: SpacingSpaceScale;
+  /** Negative bottom space around the element */
   bottom?: SpacingSpaceScale;
+  /** Negative left space around the element */
   left?: SpacingSpaceScale;
+  /** Negative right space around the element */
   right?: SpacingSpaceScale;
 }
 
 export const Bleed = ({
-  spacing,
-  horizontal,
+  horizontal = '5',
   vertical,
   top,
   bottom,
@@ -53,31 +52,27 @@ export const Bleed = ({
       return directionValues.horizontal;
     } else if (!xAxis.includes(direction) && vertical) {
       return directionValues.vertical;
-    } else {
-      return spacing;
     }
   };
 
-  const negativeMargins = {
-    top: getNegativeMargins('top'),
-    left: getNegativeMargins('left'),
-    right: getNegativeMargins('right'),
-    bottom: getNegativeMargins('bottom'),
-  } as Spacing;
+  const negativeTop = getNegativeMargins('top');
+  const negativeLeft = getNegativeMargins('left');
+  const negativeRight = getNegativeMargins('right');
+  const negativeBottom = getNegativeMargins('bottom');
 
   const style = {
-    ...(negativeMargins.bottom
-      ? {'--pc-bleed-margin-bottom': `var(--p-space-${negativeMargins.bottom})`}
-      : undefined),
-    ...(negativeMargins.left
-      ? {'--pc-bleed-margin-left': `var(--p-space-${negativeMargins.left})`}
-      : undefined),
-    ...(negativeMargins.right
-      ? {'--pc-bleed-margin-right': `var(--p-space-${negativeMargins.right})`}
-      : undefined),
-    ...(negativeMargins.top
-      ? {'--pc-bleed-margin-top': `var(--p-space-${negativeMargins.top})`}
-      : undefined),
+    '--pc-bleed-margin-bottom': negativeBottom
+      ? `var(--p-space-${negativeBottom})`
+      : undefined,
+    '--pc-bleed-margin-left': negativeLeft
+      ? `var(--p-space-${negativeLeft})`
+      : undefined,
+    '--pc-bleed-margin-right': negativeRight
+      ? `var(--p-space-${negativeRight})`
+      : undefined,
+    '--pc-bleed-margin-top': negativeTop
+      ? `var(--p-space-${negativeTop})`
+      : undefined,
   } as React.CSSProperties;
 
   return (