+++ polaris-react/src/components/Layout/components/AnnotatedSection/AnnotatedSection.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import {Heading} from '../../../Heading';
+import {Box} from '../../../Box';
+import {Text} from '../../../Text';
 import {TextContainer} from '../../../TextContainer';
 import styles from '../../Layout.scss';
 
@@ -25,11 +26,11 @@ export function AnnotatedSection({
       <div className={styles.AnnotationWrapper}>
         <div className={styles.Annotation}>
           <TextContainer>
-            <Heading id={id}>{title}</Heading>
+            <Text id={id} variant="headingMd" as="h2">
+              {title}
+            </Text>
             {descriptionMarkup && (
-              <div className={styles.AnnotationDescription}>
-                {descriptionMarkup}
-              </div>
+              <Box color="text-subdued">{descriptionMarkup}</Box>
             )}
           </TextContainer>
         </div>