+++ polaris-react/src/components/Modal/components/Dialog/Dialog.tsx
@@ -1,4 +1,4 @@
-import React, {useRef, useEffect} from 'react';
+import React, {useRef, useEffect, SetStateAction, Dispatch} from 'react';
 import {Transition, CSSTransition} from 'react-transition-group';
 import {motion} from '@shopify/polaris-tokens';
 
@@ -22,6 +22,7 @@ export interface DialogProps {
   onExited?(): void;
   in?: boolean;
   fullScreen?: boolean;
+  setClosing?: Dispatch<SetStateAction<boolean>>;
 }
 
 export function Dialog({
@@ -35,6 +36,7 @@ export function Dialog({
   small,
   limitHeight,
   fullScreen,
+  setClosing,
   ...props
 }: DialogProps) {
   const containerNode = useRef<HTMLDivElement>(null);
@@ -52,6 +54,19 @@ export function Dialog({
       !containerNode.current.contains(document.activeElement) &&
       focusFirstFocusableNode(containerNode.current);
   }, []);
+
+  const handleKeyDown = () => {
+    if (setClosing) {
+      setClosing(true);
+    }
+  };
+
+  const handleKeyUp = () => {
+    if (setClosing) {
+      setClosing(false);
+    }
+    onClose();
+  };
 
   return (
     <TransitionChild
@@ -78,7 +93,12 @@ export function Dialog({
             className={styles.Dialog}
           >
             <div className={classes}>
-              <KeypressListener keyCode={Key.Escape} handler={onClose} />
+              <KeypressListener
+                keyCode={Key.Escape}
+                keyEvent="keydown"
+                handler={handleKeyDown}
+              />
+              <KeypressListener keyCode={Key.Escape} handler={handleKeyUp} />
               {children}
             </div>
           </div>