+++ polaris-react/src/components/Modal/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {DisplayText} from '../../../DisplayText';
+import {Text} from '../../../Text';
 import {CloseButton} from '../CloseButton';
 
 import styles from './Header.scss';
@@ -8,21 +8,32 @@ import styles from './Header.scss';
 export interface HeaderProps {
   id: string;
   titleHidden: boolean;
+  closing: boolean;
   children?: React.ReactNode;
   onClose(): void;
 }
 
-export function Header({id, titleHidden, children, onClose}: HeaderProps) {
+export function Header({
+  id,
+  titleHidden,
+  closing,
+  children,
+  onClose,
+}: HeaderProps) {
   return (
     <div
       className={titleHidden || !children ? styles.titleHidden : styles.Header}
     >
       <div id={id} className={styles.Title}>
-        <DisplayText element="h2" size="small">
+        <Text as="h2" variant="headingLg">
           {children}
-        </DisplayText>
+        </Text>
       </div>
-      <CloseButton titleHidden={titleHidden} onClick={onClose} />
+      <CloseButton
+        pressed={closing}
+        titleHidden={titleHidden}
+        onClick={onClose}
+      />
     </div>
   );
 }