+++ polaris.shopify.com/pages/examples/alpha-stack-with-align.tsx
@@ -6,7 +6,7 @@ import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
 function AlphaStackWithAlignExample() {
   return (
     <Page>
-      <AlphaStack spacing="5" fullWidth>
+      <AlphaStack gap="5" fullWidth>
         <AlphaStack align="start">
           <Placeholder width="320px" label="Start" childAlign="start" />
           <Placeholder width="320px" childAlign="start" />