+++ polaris.shopify.com/pages/examples/drop-zone-nested.tsx
@@ -1,4 +1,4 @@
-import {DropZone, Stack, Thumbnail, Caption, Card} from '@shopify/polaris';
+import {DropZone, Stack, Thumbnail, Card, Text} from '@shopify/polaris';
 import {NoteMinor} from '@shopify/polaris-icons';
 import {useState, useCallback} from 'react';
 import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
@@ -29,7 +29,10 @@ function NestedDropZoneExample() {
             }
           />
           <div>
-            {file.name} <Caption>{file.size} bytes</Caption>
+            {file.name}{' '}
+            <Text variant="bodySm" as="p">
+              {file.size} bytes
+            </Text>
           </div>
         </Stack>
       ))}