+++ polaris.shopify.com/src/components/HomePage/HomePage.tsx
@@ -74,6 +74,7 @@ function HomePage({}: Props) {
               <Image
                 width={1600}
                 height={800}
+                style={{width: '100%', height: 'auto'}}
                 src="/images/home-news.png"
                 alt="A list showcasing the new font sizes in Polaris"
               />